Switch urls in countries.json to use GitHub directly
RawGit has been down a couple of times lately, so instead of pointing to
that this changes things to point directly to GitHub. This does mean
that all JSON and CSV urls will have the content type text/plain (as
that's what GitHub serves them up with) but we can document that and
tell people to use RawGit if they need the correct content type.